+++ src/features/home/Home.jsx
@@ -2,9 +2,8 @@ import React from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
 import { ArrowRight, BookOpen, Clock3, Search, Sparkles } from "lucide-react";
 import Navbar from "../../shared/components/Navbar";
-import HeroSection from "./HeroSection";
 import Footer from "../../shared/components/Footer";
-import FeaturedToolsSection from "./FeaturedToolsSection";
+import { FeaturedToolsSection, HeroSection } from "./components";
 import homeData from "./HomeData";
 import { useTheme } from "../../shared/context/ThemeContext";
 import coreTopics from "../../shared/data/coreTopics";
