+++ src/features/home/components/FeaturedToolsSection.jsx
