+++ src/features/home/components/index.js
@@ -1,0 +1,2 @@
+export { HeroSection } from './HeroSection';
+export { FeaturedToolsSection } from './FeaturedToolsSection';
